requirements.txt
# Borsdata API Fetcher & ML Pipeline Requirements
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
